VM: add division, bitwise, shift ops; lift the 64-global limit
New integer opcodes, end to end (VM, IntOp, constant folding, emit,
disasm, Scheme and Fleche): INT_DIV, INT_MOD (Rocq Nat convention:
x / 0 = 0, x mod 0 = x, truncating), INT_SUB_SAT (Nat.sub), INT_LE,
INT_AND/OR/XOR, INT_SHL and logical INT_SHR (shift amounts outside
0..24 give 0). The semantics live in encore_vm::int, shared by the VM
and the constant folder so they agree bit for bit.

Globals: the VM no longer has a fixed [_; 64] table. load() carves one
word per global from the top of the heap buffer, and fails with
GlobalsOverflow when they do not fit. Program reads globals from its
source instead of copying into a fixed array. The asm IR uses u16
global indices, emitted as GLOBAL_W for indices >= 256. The global-name
metadata index is now u16.

Hard limits are reported instead of wrapping: CompileError for too
many globals, code over 64 KB, over-long byte literals, too many
captures and too-wide constructors; a parse error past 256
constructor tags. compile_module now returns a Result.

Closes #20

+++ crates/encore_compiler/src/ir/asm.rs
@@ -39,7 +39,7 @@ pub struct Case {
 pub enum Val {
     Reg(Reg),
     Capture(u8),
-    Global(u8),
+    Global(u16),
     ContLam(ContLam),
     Ctor(Tag, Vec<Reg>),
     Field(Reg, u8),
@@ -60,7 +60,7 @@ pub enum Expr {
 
 #[derive(Debug)]
 pub struct Define {
-    pub global: u8,
+    pub global: u16,
     pub body: Expr,
 }
